Upgraded TextValidateField
